[scaling] feat: builder service (#999) - only test refactor
* feat: builder init

* feat: build route

* feat: builder service

* feat: binary and docker image

* refactor: clippy suggestions

* refactor: move test helpers to common crate

* refactor: trim deployer-alpha

* refactor: get name from nbuild

* refactor: .crate to .tar.gz

* refactor: Cmd to Entrypoint
