+++ common-tests/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "shuttle-common-tests"
+version.workspace = true
+edition.workspace = true
+license.workspace = true
+repository.workspace = true
+publish = false
+
+[dependencies]
+hyper = { workspace = true }
+shuttle-common = { workspace = true }
+tower = { workspace = true }
